Revert "Make builders abstract to allow for testing"
This reverts commit 4176fe6d63736b01adcac926910f5890b8f517c3.
