+++ firebase-firestore/src/commonMain/kotlin/dev/gitlive/firebase/firestore/FieldsAndValuesBuilder.kt
@@ -1,10 +1,12 @@
 package dev.gitlive.firebase.firestore
 
 import dev.gitlive.firebase.EncodeSettings
+import dev.gitlive.firebase.firestore.internal.FieldAndValue
 import kotlinx.serialization.SerializationStrategy
 
-public abstract class FieldsAndValuesBuilder {
+public class FieldsAndValuesBuilder internal constructor() {
 
+    internal val fieldAndValues: MutableList<FieldAndValue> = mutableListOf()
     public var buildSettings: EncodeSettings.Builder.() -> Unit = {
         encodeDefaults = true
     }
@@ -26,8 +28,12 @@ public abstract class FieldsAndValuesBuilder {
     }
 
     @PublishedApi
-    internal abstract fun String.toEncoded(encodedValue: Any?)
+    internal fun String.toEncoded(encodedValue: Any?) {
+        fieldAndValues += FieldAndValue.WithStringField(this, encodedValue)
+    }
 
     @PublishedApi
-    internal abstract fun FieldPath.toEncoded(encodedValue: Any?)
+    internal fun FieldPath.toEncoded(encodedValue: Any?) {
+        fieldAndValues += FieldAndValue.WithFieldPath(this, encodedValue)
+    }
 }
